master version

+++ HelloWorld.java
@@ -1,7 +1,7 @@
 public class HelloWorld {
 
     public static void main(String[] args) {
-		private float defense;
+		private double defense;
         System.out.println("Hello World!");
     }
